requirements.txt
beautifulsoup4==4.8.2
bs4==0.0.1
certifi==2019.11.28
chardet==3.0.4
idna==2.8
numpy==1.18.1
pandas==1.0.1
Pillow==7.0.0
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
selenium==3.141.0
six==1.14.0
soupsieve==1.9.5
urllib3==1.25.8
rich==6.1.1
xlrd

# Installer
pyinstaller

chromedriver-autoinstaller
dnspython
tldextract
